Removes unnecessary boxing, hazardous BigDecimal uses and other tripwires
The return value of Customer.reduceBalance() is never used and therefore removed.

The mapperFactory in DtoMapper is only used for initialisation. Therefore it was moved to the initialisation block directly.

+++ src/orderdomainWAR/src/main/java/de/novatec/showcase/order/ejb/session/OrderSession.java
@@ -73,7 +73,7 @@ public class OrderSession implements OrderSessionLocal {
 		TypedQuery<Long> orderCount = em.createNamedQuery(Order.COUNT_BY_CUSTOMER, Long.class);
 		orderCount.setParameter("id", customerId);
 		List<Long> elementList = orderCount.getResultList();
-		return elementList != null ? elementList.get(0).longValue() : Long.valueOf(0).longValue();
+		return elementList != null ? elementList.get(0) : Long.valueOf(0);
 	}
 
 	@Override
